Add tutorials 8 and 10
Tutorial 8 adds some initial DOM manipulation, on the shopping form,
adding a (sort of) tooltip that pops up when you hover on the
"calculate" button. In the process, we introduce `hiccups`, a subset
of `hiccup` that has been ported to clojurescript.

Tutorial 10 (we skip 9, which is on basic ajax and which we'll cover in the
next little project) adds both domina events and some DOM manipulation
to give nicer error messages and stronger validation in the browser.
The regexps we use are interesting, I actually learned a bit by studying
the password one and figuring out what the ?= (lookahead) syntax was
doing.

I'd like to point out that as of this commit, the clojurescript tutorial
is still 100% compatible with using a python webserver. Starting in the
next part of the project, we will be using a clojure webserver for
functionality and this will no longer be the case.