Add a json POST endpoint to the server
This actually took me a long while to figure out; as I discuss
in the README, I should definitely read a book or some clojure
tutorials before I ever try to build anything more substantial
than this trivial server. It's not that the tools aren't good,
they're just not obvious for me to figure out at this point.

I'm not using the POST endpoint on the client yet, but I
added a curl script you can use to verify that it works.

By the way: a POST request isn't really right here, since
the calculate function isn't modifying a resource; I could
have just used a GET, and figured out how to parse the params.
But I expect that a similar real app would likely need a
POST so I decided to figure out how to implement one.