Add the client-side use of the POST endpoint
For the server-side, I decided to do a vanilla json api rather
than shoreleave. Similarly, on the client side I went with
cljs-ajax, which is a vanilla ajax library that can handle
several transfer formats, including json.

One awesome advantage of cljs-ajax is that it is supposed to
be portable to clojure, so I can likely use it in a command-line
tool or microservice (although it may not be the best choice
for that in most cases)