Add some simple ajax and api for the login form
This corresponds loosely to Tutorial 11, although I wasn't interested
in the part of the tutorial that involved using html5, since I want
to write only json apis.

In the process, I did a decent bit of refactoring of the client-side
code to get `valid-email?` and `valid-password?`, which for now are
copy-pasted to the server-side code.

The next step (Tutorial 12) will be to use a cljc file to factor
out the common logic.