Add testing to boot
I was unfortunately only able to get the clojure-side tests
working, not the clojurescript-side ones. But on the bright side,
clojure now seems to be reloading in the `serve` task (I have no
idea what changed), and also in the `test` task.

Also this finishes Tutorial 15, which is probably the last one
I'll do for now aside from maybe the react porting examples... it's
time to find an actual project to work on before I dig anymore into
modern-cljs, especially given that it's unlear whether `boot` or
`lien` is a better choice for my work.